pyproject.toml
[project]
name = "ecommerce-datawarehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.10.15",
    "dbt-postgres>=1.9.1",
    "kaggle>=1.8.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
